put, assign & remove students & lecturer api added

+++ route/course_routes.js
@@ -2,15 +2,69 @@ const router = require("express").Router();
 
 const auth = require("../middleware/auth");
 const CourseService = require("../service/course_service");
-
+const Error = require("../model/error");
 const courseService = new CourseService();
+
+router.post("/", auth, async (req, res) => {
+	try {
+		const type = req.user.type;
+		if (type !== 0) throw new Error("Unauthorized", 401);
+		const result = await courseService.createCourse(req.body);
+		return res.status(200).json({ status: "OK", message: result });
+	} catch (err) {
+		return res.status(err.statusCode).json(err);
+	}
+});
+
+router.post("/assign/student", auth, async (req, res) => {
+	try {
+		const type = req.user.type;
+		if (type !== 0) throw new Error("Unauthorized", 401);
+		const result = await courseService.assignStudentToCourse(req.body);
+		return res.status(200).json({ status: "OK", message: result });
+	} catch (err) {
+		return res.status(err.statusCode).json(err);
+	}
+});
+
+router.post("/assign/lecturer", auth, async (req, res) => {
+	try {
+		const type = req.user.type;
+		if (type !== 0) throw new Error("Unauthorized", 401);
+		const result = await courseService.assignLecturerToCourse(req.body);
+		return res.status(200).json({ status: "OK", message: result });
+	} catch (err) {
+		return res.status(err.statusCode).json(err);
+	}
+});
 
 router.get("/", auth, async (req, res) => {
 	try {
+		const courseID = req.query.id;
+		if (courseID != null) {
+			const result = await courseService.getCourseByID(courseID);
+			return res.status(200).json({ status: "OK", course: result });
+		}
 		const semester = req.query.semester;
-		const courses = await courseService.getCourseBySemseter(semester, req.user);
+		const courses = await courseService.getCourseBySemester(semester, req.user);
 		return res.status(200).json({ status: "OK", courses: courses });
 	} catch (err) {
+		return res.status(err.statusCode).json(err);
+	}
+});
+
+router.put("/", auth, async (req, res) => {
+	try {
+		const type = req.user.type;
+		if (type !== 0) throw new Error("Unauthorized", 401);
+
+		const courseID = req.query.id;
+		if (courseID == null) throw new Error("Bad request", 401);
+		console.log("here");
+		const result = await courseService.updateCourseByID(courseID, req.body);
+		return res.status(200).json({ status: "OK", message: result });
+	} catch (err) {
+		console.log(err.statusCode);
 		return res.status(err.statusCode).json(err);
 	}
 });
@@ -18,12 +72,37 @@ router.get("/", auth, async (req, res) => {
 router.delete("/", auth, async (req, res) => {
 	try {
 		const type = req.user.type;
+		if (type !== 0) throw new Error("Unauthorized", 401);
 		const id = req.query.id;
 		const response = await courseService.deleteCourseByID(id, type);
 		return res.status(200).json({
 			status: "OK",
 			message: response + " course deleted successfully",
 		});
+	} catch (err) {
+		return res.status(err.statusCode).json(err);
+	}
+});
+
+router.delete("/remove/student", auth, async (req, res) => {
+	try {
+		const type = req.user.type;
+		if (type !== 0) throw new Error("Unauthorized", 401);
+		const result = await courseService.removeStudentFromCourse(req.body);
+		if (result == 0) throw new Error("Not found", 404);
+		return res.status(200).json({ status: "OK", message: result });
+	} catch (err) {
+		return res.status(err.statusCode).json(err);
+	}
+});
+
+router.delete("/remove/lecturer", auth, async (req, res) => {
+	try {
+		const type = req.user.type;
+		if (type !== 0) throw new Error("Unauthorized", 401);
+		const result = await courseService.removeLecturerFromCourse(req.body);
+		if (result == 0) throw new Error("Not found", 404);
+		return res.status(200).json({ status: "OK", message: result });
 	} catch (err) {
 		return res.status(err.statusCode).json(err);
 	}
